cleanup
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@2975 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-proxy/src/main/java/org/sonatype/nexus/proxy/repository/AbstractProxyRepository.java
@@ -734,7 +734,7 @@ public abstract class AbstractProxyRepository
         // else check age
         else
         {
-            return ( ( System.currentTimeMillis() - item.getRemoteChecked() ) > ( (long) maxAge * 60L * 1000L ) );
+            return ( ( System.currentTimeMillis() - item.getRemoteChecked() ) > ( maxAge * 60L * 1000L ) );
         }
     }
